refactor(activerecord): move connection-handler, schema-dumper, join-dependency, fixtures
Move 4 more files to Rails-canonical locations:
- connection-handler.ts → connection-adapters/abstract/connection-handler.ts
- schema-dumper.ts → connection-adapters/abstract/schema-dumper.ts
- join-dependency.ts → associations/join-dependency/join-association.ts
- fixtures.ts → fixture-set/file.ts

All imports updated to reference new locations directly.
api:compare: 47/597 (7.9%), 15 misplaced (was 19).
