+++ packages/activerecord/src/connection-adapters/abstract/connection-handler.ts
@@ -4,11 +4,11 @@
  * Mirrors: ActiveRecord::ConnectionAdapters::ConnectionHandler
  */
 
-import { ConnectionPool } from "./connection-pool.js";
-import { DatabaseConfig } from "./database-configurations/database-config.js";
-import { HashConfig } from "./database-configurations/hash-config.js";
-import { DatabaseConfigurations } from "./database-configurations/connection-url-resolver.js";
-import type { DatabaseAdapter } from "./adapter.js";
+import { ConnectionPool } from "../../connection-pool.js";
+import { DatabaseConfig } from "../../database-configurations/database-config.js";
+import { HashConfig } from "../../database-configurations/hash-config.js";
+import { DatabaseConfigurations } from "../../database-configurations/connection-url-resolver.js";
+import type { DatabaseAdapter } from "../../adapter.js";
 
 export class ConnectionHandler {
   private _pools = new Map<string, ConnectionPool>();
