+++ packages/activerecord/src/connection-adapters/connection-handler.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach } from "vitest";
-import { ConnectionHandler } from "../connection-handler.js";
+import { ConnectionHandler } from "./abstract/connection-handler.js";
 import { HashConfig } from "../database-configurations/hash-config.js";
 import { DatabaseConfigurations } from "../database-configurations/connection-url-resolver.js";
 import { createTestAdapter } from "../test-adapter.js";
